Rewrite all the macros to support custom scalar value representations
This commit changes all provided macros to support the introduced
scalar value abstraction. It is now possible to specify if a certain
implementation should be based on a specific scalar value
representation or be generic about the exact representation. All
macros now defaulting to the `DefaultScalarValue` type provided by
juniper if no scalar value representation is specified. This is done
with usability and backwards compatibility in mind.

To implement this feature more or less the whole macro code is
rewritten. Now there is a clear separation between parsing the custom
dsl and expanding the final impls. Also the new code tries to share
much of the implementation between the different macros.