Fix a bug regarding to input value casting in custom scalars
This fix introduces a new way to convert a scalar value into one of
the 4 standardized scalar values (Int, Float, String, Bool). Using
this abstraction it is possible to do certain needed conversations in
custom scalar value code, for example that one that converts a integer
value to a floating point value.