Give `IntoFieldError::into_field_error` access to the context
Fixes https://github.com/graphql-rust/juniper/issues/391

`IntoFieldError` is useful for sending errors to error tracking services
such as Rollbar or Bugsnag. When you do that it is often useful to
include data accessible via the context such as data about the current
user.

Getting access to the context inside `IntoFieldError` previously required
[wrapping your error][] which lead to a lot of boilerplate. This changes that
so `IntoFieldError` gets passed a reference to the context.

This is a breaking change but I think the small bit of breakage is worth it. I
doubt users have more than one implementation of `IntoFieldError`.

[wrapping your error]: https://github.com/graphql-rust/juniper/issues/391