Add futures dependency
Fixes https://github.com/graphql-rust/juniper/issues/534.

I am not ure if the best way is to have folks bring their own futures for the
derives / macros or to pull it in as a dep explicitly like I have done here.

Because people are filing issues, we will do it this way and see what other issues arise.