Remove duplicate link to Hyper integration
The `Adding a Server` page contains a duplicate link to `Hyper`. The first of them can cause the reader to accidentally skip the `Rocket` and `Iron` integration pages because their pages are before the one for `Hyper` in the reading order. 

Remove the first link so that the list of official server integrations matches the page order.