Match the list of integrations to the page order (#548)
* Remove duplicate link to Hyper integration

The `Adding a Server` page contains a duplicate link to `Hyper`. The first of them can cause the reader to accidentally skip the `Rocket` and `Iron` integration pages because their pages are before the one for `Hyper` in the reading order. 

Remove the first link so that the list of official server integrations matches the page order.

* Match the list of integrations to the page order

In the book, the pages for the official integrations appear in the order: `Warp`, `Rocket`, `Iron`, `Hyper`, but on the `Official Server Integrations` page they are listed with `Hyper` first. This can cause readers to accidentally skip the other three integrations if they click on the link for `Hyper` first. 

Move `Hyper` to the bottom of the list to match the reading order of the book.

* Remove unnecessary blank line