refactor: remove ScalarValue type parameter
The type parameter S for ScalarValue is used to open up a degree of freedom in the parsing of leaf scalar types. However because of that it is threaded through the entire codebase. This commit removes the type parameter in all possible places with the following in mind:
* Fix the S parameter to DefaultScalarValue where necessary (mainly in Value which occurs on the return path after resolution)
* Remove the parameter everywhere else

This commit effectively disables the feature. I'm committing here mainly because now it compiles so I want a safe point.

Next we need to figure out how to re-add the ability to control the leaf value parsing